requirements.txt
lxml
spacy
faiss-cpu
streamlit
openai
dotenv
pydantic==1.10.9
tiktoken
matplotlib
sentence-transformers
streamlit-extras
beautifulsoup4
tqdm
requests
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-google-vertexai
google-generativeai
google-ai-generativelanguage
google-cloud-discoveryengine
google-cloud-aiplatform[reasoningengine,langchain]